refactor: move make_coincident to data_handling
