+++ utilities/data_handling.py
@@ -6,8 +6,43 @@ from bpy.types import Scene, Context
 
 from ..model.types import SlvsGenericEntity, SlvsSketch, GenericConstraint
 from ..model.utilities import update_pointers
+from ..model.workplane import SlvsWorkplane
+from ..model.categories import LINE, CURVE, POINT
 
 logger = logging.getLogger(__name__)
+
+
+# NOTE: When tweaking, it's necessary to constrain a point that is only temporary available
+# and has no SlvsPoint representation
+def make_coincident(solvesys, point_handle, e2, wp, group, entity_type=None):
+    func = None
+    set_wp = False
+
+    if entity_type:
+        handle = e2
+    else:
+        entity_type = type(e2)
+        handle = e2.py_data
+
+    if entity_type in LINE:
+        func = solvesys.addPointOnLine
+        set_wp = True
+    elif entity_type in CURVE:
+        func = solvesys.addPointOnCircle
+    elif entity_type == SlvsWorkplane:
+        func = solvesys.addPointInPlane
+    elif entity_type in POINT:
+        func = solvesys.addPointsCoincident
+        set_wp = True
+
+    kwargs = {
+        "group": group,
+    }
+
+    if set_wp:
+        kwargs["wrkpln"] = wp
+
+    return func(point_handle, handle, **kwargs)
 
 
 def to_list(value):
